Lambdas_functions/pyproject.toml
[project]
name = "lambdas"
version = "0.1.0"
description = "Here we can find the code pushed to the different lambdas used in AWS"
authors = [
    {name = "Amine.B"},
    {name = "Samy.A"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "boto3 (>=1.41.1,<2.0.0)",
    "awscli (>=1.43.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "llama-index (>=0.14.8,<0.15.0)",
    "pymupdf-layout (>=1.26.6,<2.0.0)",
    "pymupdf4llm (>=0.2.3,<0.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
